requirements.txt
pillow~=10.3.0